requirements.txt
absl-py==0.3.0
alabaster==0.7.11
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
Babel==2.6.0
backcall==0.1.0
bleach==2.1.3
bokeh==0.13.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.3
cryptography==2.3.1
cycler==0.10.0
Cython==0.28.5
dask==0.18.2
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
html5lib==1.0.1
idna==2.7
imagesize==1.0.0
imgaug==0.2.6
imutils==0.5.1
ipykernel==4.8.2
ipyparallel==6.2.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
isort==4.3.4
jedi==0.12.1
jeepney==0.4
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
keyring==15.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nose==1.3.7
notebook==5.5.0
numpy==1.14.5
numpydoc==0.8.0
opencv-contrib-python==3.4.3.18
opencv-python==3.4.2.17
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.7
ptyprocess==0.6.0
pycocotools==2.0.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.0
PyQt5==5.9.2
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.0.0
QtAwesome==0.5.1
qtconsole==4.4.1
QtPy==1.5.1
requests==2.19.1
rope==0.11.0
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
SecretStorage==3.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
sip==4.19.8
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.7.7
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0.2
tqdm==4.24.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
wrapt==1.10.11
